@@ -43,21 +43,22 @@
     "key-mirror": "^1.0.1",
     "lodash": "^4.17.4",
     "markdown-loader": "^2.0.0",
-    "node-sass": "^4.5.0",
+    "node-sass": "^4.5.3",
     "postcss-loader": "^1.3.2",
-    "react": "^15.4.2",
-    "react-addons-shallow-compare": "^15.4.2",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-addons-shallow-compare": "^15.6.2",
     "react-autobind": "^1.0.6",
-    "react-dom": "^15.4.2",
-    "react-google-maps": "^6.3.0",
-    "react-hot-loader": "next",
+    "react-dom": "^16.0.0",
+    "react-google-maps": "^9.0.1",
+    "react-hot-loader": "^3.1.1",
     "react-localization": "^0.0.13",
-    "react-markdown": "^2.4.6",
-    "react-redux": "^5.0.3",
-    "react-router": "^4.0.0",
-    "react-router-dom": "^4.0.0",
+    "react-markdown": "^2.5.0",
+    "react-redux": "^5.0.6",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "redux": "^3.6.0",
-    "redux-form": "^6.5.0",
+    "redux-form": "^7.1.1",
     "redux-localstorage": "^0.4.1",
     "require-dir": "^0.3.2",
     "reselect": "^3.0.1",
